refactor: split logic into service and controller
