+++ src/utilities/file.service.ts
@@ -1,7 +1,10 @@
 import { readdir, readFile } from 'fs/promises';
 import { HttpException, HttpStatus, Injectable, Logger, StreamableFile } from '@nestjs/common';
-import { PathLike } from 'fs';
+import { existsSync, mkdirSync, PathLike, readdirSync, rmSync, readFileSync } from 'fs';
 import * as xml2js from 'xml2js';
+import { getDocument } from 'pdfjs-dist/legacy/build/pdf';
+
+const pdf = require('pdf-poppler');
 
 @Injectable()
 export class FileService {
@@ -57,6 +60,54 @@ export class FileService {
             const message = `Error retrieving file: ${fileName} in folder:${directory}`;
             this.logger.error(message, err);
             throw new HttpException(message, HttpStatus.NOT_FOUND);
+        }
+    }
+
+    async getNumberOfPdfPages(fileName: string): Promise<number> {
+        const retrievedFile = readFileSync(`resources\\pdfs\\${fileName}`);
+
+        return getDocument({ data: retrievedFile }).promise.then((document) => document.numPages);
+    }
+
+    async getConvertedPdfFile(fileName: string, pageNumber: number): Promise<StreamableFile> {
+        const pagesFolder = `${process.cwd()}\\out\\${fileName.replace('.pdf', '')}`;
+
+        if (!existsSync(pagesFolder)) {
+            mkdirSync(pagesFolder);
+        }
+
+        const options = {
+            format: 'png',
+            out_dir: pagesFolder,
+            out_prefix: 'out',
+            page: pageNumber,
+            scale: 2048,
+        };
+
+        try {
+            return pdf.convert(`${process.cwd()}\\resources\\pdfs\\${fileName}`, options).then(() => {
+                const fileList = readdirSync(pagesFolder, { withFileTypes: true });
+
+                if (!fileList.length) {
+                    throw new Error('No files found in the output folder');
+                }
+                const pageNumberLength = fileList[0].name.replace('out-', '').replace('.png', '').length;
+                const expectedFilePath = `${pagesFolder}\\out-${pageNumber.toString().padStart(pageNumberLength, '0')}.png`;
+
+                if (existsSync(expectedFilePath)) {
+                    const file = new StreamableFile(readFileSync(expectedFilePath));
+
+                    rmSync(expectedFilePath);
+
+                    return file;
+                }
+
+                throw new Error('No file found in the output folder');
+            });
+        } catch (err) {
+            const message = `Error converting PDF to PNG: ${fileName}`;
+            this.logger.log(message, err);
+            throw new HttpException(message, HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
